exported stakeDOUGH function into staking-helper
